sideBar complete
