+++ packages/client/src/components/drawerBar/drawerBarData.tsx
@@ -1,0 +1,15 @@
+import Main from './icons/Main'
+import ProfileIcon from './icons/Profile'
+import WinnerIcon from './icons/Winner'
+import ForumIcon from './icons/Forum'
+import GameIcon from './icons/Game'
+
+export const menuData = [
+  { text: 'Главная', icon: <Main /> },
+  { text: 'Профиль', icon: <ProfileIcon /> },
+  { text: 'Лидеры', icon: <WinnerIcon /> },
+  { text: 'Форум', icon: <ForumIcon /> },
+  { text: 'Начать игру', icon: <GameIcon /> },
+]
+
+export const drawerWidth = 355
