+++ packages/client/src/components/drawerBar/icons/AvatarIcon.tsx
@@ -1,0 +1,23 @@
+import { SvgIcon, SvgIconProps } from '@mui/material'
+
+interface AvatarIconProps extends SvgIconProps {
+  open?: boolean
+}
+
+export default function AvatarIcon(props: AvatarIconProps) {
+  const viewBox = props.open ? '-0 0 70 70' : '-10 -10 90 90'
+  return (
+    <SvgIcon {...props} viewBox={viewBox}>
+      <path
+        d="M35 7C19.5544 7 7 19.5965 7 35.1761C7 39.5943 8.00806 43.7651 9.80151 47.4775C10.6424 49.218 9.91303 51.3107 8.1725 52.1515C6.43197 52.9924 4.33934 52.263 3.49849 50.5225C1.25579 45.8802 0 40.6704 0 35.1761C0 15.7673 15.6517 0 35 0C54.3483 0 70 15.7673 70 35.1761C70 40.6704 68.7442 45.8802 66.5015 50.5225C65.6607 52.263 63.568 52.9924 61.8275 52.1515C60.087 51.3107 59.3576 49.218 60.1985 47.4775C61.9919 43.7651 63 39.5943 63 35.1761C63 19.5965 50.4456 7 35 7Z"
+        fill="#1E515D"
+      />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M43.7902 50.1355C48.9988 47.1039 52.5 41.4608 52.5 35C52.5 25.335 44.665 17.5 35 17.5C25.335 17.5 17.5 25.335 17.5 35C17.5 41.4608 21.0012 47.1039 26.2098 50.1355C20.8525 51.5872 16.754 54.2324 13.7813 56.8912C11.5839 58.8566 10.0055 60.8229 8.96817 62.3116C8.44821 63.0579 8.0601 63.6901 7.79464 64.1506C7.66178 64.3811 7.55928 64.5693 7.48599 64.708C7.44934 64.7774 7.41996 64.8345 7.39772 64.8784L7.36968 64.9343L7.35963 64.9546L7.35559 64.9629L7.35382 64.9665C7.353 64.9682 7.35221 64.9698 10.5 66.5L7.35221 64.9698C6.50712 66.7083 7.23135 68.8027 8.96983 69.6478C10.7036 70.4906 12.7913 69.7726 13.6409 68.0443L13.6429 68.0403L13.6755 67.9776C13.7106 67.9112 13.7715 67.7986 13.8593 67.6463C14.0352 67.3412 14.317 66.8796 14.7115 66.3134C15.5032 65.1771 16.7321 63.6434 18.4479 62.1088C21.846 59.0695 27.1697 56 35 56C42.8303 56 48.154 59.0695 51.5521 62.1088C53.2679 63.6434 54.4968 65.1771 55.2885 66.3134C55.683 66.8796 55.9648 67.3412 56.1407 67.6463C56.2285 67.7986 56.2894 67.9112 56.3245 67.9776L56.3571 68.0403L56.3591 68.0442C57.2086 69.7726 59.2964 70.4906 61.0302 69.6478C62.7686 68.8027 63.4929 66.7083 62.6478 64.9698L59.5 66.5C62.6478 64.9698 62.647 64.9682 62.6462 64.9665L62.6404 64.9546L62.6303 64.9343L62.6023 64.8784C62.58 64.8345 62.5507 64.7774 62.514 64.708C62.4407 64.5693 62.3382 64.3811 62.2054 64.1506C61.9399 63.6901 61.5518 63.0579 61.0318 62.3116C59.9945 60.8229 58.4161 58.8566 56.2187 56.8912C53.246 54.2324 49.1475 51.5872 43.7902 50.1355ZM35 45.5C40.799 45.5 45.5 40.799 45.5 35C45.5 29.201 40.799 24.5 35 24.5C29.201 24.5 24.5 29.201 24.5 35C24.5 40.799 29.201 45.5 35 45.5Z"
+        fill="#1E515D"
+      />
+    </SvgIcon>
+  )
+}
