+++ packages/client/src/components/drawerBar/icons/LeftArrow.tsx
