+++ packages/client/src/components/drawerBar/icons/Main.tsx
