+++ packages/client/src/components/drawerBar/icons/RightArrow.tsx
