+++ packages/client/src/components/drawerBar/icons/Winner.tsx
