+++ packages/client/src/components/drawerBar/index.tsx
@@ -1,0 +1,1 @@
+export { DrawerBar } from './drawerBar'
