--- packages/client/src/components/sideBar/index.tsx
@@ -1,1 +1,0 @@
-export { NavBar } from './sideBar'
